requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
python-multipart==0.0.20
httpx==0.27.2
tenacity==9.0.0
openai==1.59.7
python-dateutil==2.9.0.post0
jinja2==3.1.5
